requirements.txt
django>=3.2.0, <3.3.0
redis
django-tempus-dominus==5.1.2.17
Pillow==9.5.0
dj-database-url==2.0.0
django-tempus-dominus==5.1.2.17
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6